requirements.txt
absl-py==1.4.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blis==0.7.10
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
confection==0.1.0
cymem==2.0.7
dataclasses-json==0.5.13
dill==0.3.7
distro==1.9.0
docstring-parser==0.15
elastic-transport==8.4.0
elasticsearch==8.8.2
faiss-cpu==1.7.4
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h11==0.14.0
h5py==3.9.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.16.4
idna==3.4
instructor==0.4.7
Jinja2==3.1.2
joblib==1.3.1
keras==2.13.1
-e git+https://github.com/langchain-ai/langchain.git@af788b7cf099fd8617a8a5b2c894522c6ca9d384#egg=langchain&subdirectory=libs/langchain
langcodes==3.3.0
langsmith==0.0.14
libclang==16.0.6
manifest-ml==0.0.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==1.8.0
openapi-schema-pydantic==1.2.4
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
pathy==0.10.2
preshed==3.0.8
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyMySQL==1.1.1
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
safetensors==0.3.1
shell_gpt==1.1.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.49
sqlitedict==2.1.0
srsly==2.4.7
sympy==1.12
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
thinc==8.1.10
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.31.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.16
wasabi==1.1.2
Werkzeug==2.3.6
wikipedia==1.4.0
wrapt==1.15.0
yarl==1.9.2
youtube-transcript-api==0.6.1
